NetcdfCFWriter : set large file if needed   FileWriter: allow largeFiles. also use much larger buffer   H5header: message sizes are unsigned

+++ ui/src/main/java/ucar/nc2/ui/Hdf5Table.java
@@ -241,7 +241,7 @@ public class Hdf5Table extends JPanel {
       return m.getName();
     }
 
-    public short getSize() {
+    public int getSize() {
       return m.getSize();
     }
